Short-term brought the Debian patchsets in...
Keep them around for a working crib-sheet now that I've got the key
ones we wanted in the mix patched in.  To be sure, there's more and
there might be some of the scripting brought in.  the init scripting
is more per-distro, but we should have some basics like a failback
as was done with patch 0008, which emulates sysv runlevel 5's behavior.