Moved current compilable version to David's suggested...
Merged the to this point cmake-dsugar branch to the mods branch (Since
all of what was suggested has either agreed upon or moved to what was
agred upon.

Cleaned up some CVS/SVN-isms out of the build that caused a problem
with VERSION being defined "properly" in config.h.in.  Still work
on cleanup left to be done, but it compiles now with the right
Version info.