First pass complete.  All main binaries building clean-ish.
It builds roughly 64-bit clean on Mint.  With a few more cleanups
this looks promising.  Going through the old makefile to make sure
I've not missed any of the crucial binaries.  Once done we'll
do a bit of a re-org to make CMake a bit cleaner and easier to
follow (One huge honkin' src/ dir for all of this is...painful)