First pass at a restructure that is easier to work with.
Now that we've got CMake building the stuff mostly as intended, we
now need to start to re-structure the repository and source code
so that someone can more easily work on the thing.  It was a monolithic
blob of source files that had no concept of the apps or library
functions- or if you even NEEDED them with more modern OSes.