First pass of CMake restructure complete.
Source for apps is separated from library components.

Library components are in ONE .a file (one, three, same size...) having
them as separate subdirs and aggregated in the upper level CMakeLists
file as a single .a comprising of three different logical CMake Object
library sets, makes it easier to specify app linkage and to figure
out what is where for audit, cleanup, and cruft removal in the next
cleaning passes.