Added what I believe will let Dave's FreeBSD changes work
Builds on my system.  Checks for the presence of utmpx.h AND utmp.h
If both are there, the preference is for utmp.h's path because there
is no guarantees that it will do anything other than cowpile on a
system.  This presumption should work as FreeBSD _*SEEMS*_ to be
the only one that does this and they have all the guts from the old
paths to build.  Will need to probably revisit with a FreeBSD
system set up in a VM.