Cruft removal pass 1...
Removed some pure trash.  errno.h and strerror is the canonical
interfaces to POSIX systems anyhow- and since we're supporting
pretty much only POSIX systems, removed the string junk since
the system versions provide the right info, ALWAYS- and used the
limited wrapper for now (May be subject to another refactor pass
down the line in the first security audit).

Making your own is NOT always a good idea- esp. if you're not Bernstein.

If it is a Unix-like that doesn't support these interfaces properly
we DO NOT SUPPORT IT AT THIS TIME.  It either has sound, secure
interfaces to the likes of these things or not.  If not, supporting
it opens up Heartbleed type fails.