Clean up a bit of iffy code...
In this case, we don't want fall-through anything in our
codebase here.  It needs to be explicit, clearly deterministic
behavior in everything.  Found a fall-through when I was
trying to sort out what timeout is "default" which isn't documented
but clearly in play with the current codebase.  Fixing this
right now.